requirements.txt
numpy==1.22.3
pandas==1.4.1
tensorflow==2.8.0
keras==2.8.0
scikit-learn==1.0.2
seaborn==0.11.2
opencv-python==4.5.5.62
Pillow==9.0.1
SimpleITK==2.1.1
matplotlib==3.5.1
